rebase to match current state of leading branch and fix tests accordingly
